+++ scripts/dump-clang-ast.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-clang -Xclang -ast-dump -fsyntax-only $1
+clang -Wdocumentation -fparse-all-comments -Xclang -ast-dump -fsyntax-only $1